Changes to strlcat strlcpy strtrim

+++ ft_strlcpy.c
@@ -6,16 +6,16 @@
 /*   By: ademurge <ademurge@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/03 18:10:09 by ademurge          #+#    #+#             */
-/*   Updated: 2022/04/11 13:33:40 by ademurge         ###   ########.fr       */
+/*   Updated: 2022/04/13 14:36:54 by ademurge         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcpy(char *dest, const char *src, size_t size)
+size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	c;
+	size_t	i;
+	size_t	c;
 
 	i = -1;
 	c = 0;
